chore: move service reference into service file
