+++ internal/swarm/service.go
@@ -2,6 +2,7 @@ package swarm
 
 import (
 	"errors"
+	"fmt"
 	"time"
 )
 
@@ -110,4 +111,35 @@ type ServiceLogsOptions struct {
 	Since *time.Time
 	// Until defines upper bound for log timestamps.
 	Until *time.Time
+}
+
+type ServiceReference struct {
+	stackName   string
+	serviceName string
+
+	fullName string
+}
+
+// NewServiceReference creates a service reference from stack and service names.
+func NewServiceReference(stackName, serviceName string) ServiceReference {
+	return ServiceReference{
+		stackName:   stackName,
+		serviceName: serviceName,
+		fullName:    fmt.Sprintf("%s_%s", stackName, serviceName),
+	}
+}
+
+// StackName returns stack name part.
+func (r ServiceReference) StackName() string {
+	return r.stackName
+}
+
+// ServiceName returns service name part.
+func (r ServiceReference) ServiceName() string {
+	return r.serviceName
+}
+
+// Name returns full service name as "<stack>_<service>".
+func (r ServiceReference) Name() string {
+	return r.fullName
 }
